chore: rearrange hooks
